Middleware права на раздел финансы

+++ backend/database/migrations/2025_03_06_093302_create_finance_projects_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->decimal('price', 14, 2);
             $table->string('comment', 2500)->nullable();
-            $table->integer('status_id');
+            $table->unsignedBigInteger('status_id');
             $table->uuid('project_id');
             $table->uuid('coordinator_id');
             $table->uuid('lead_id');
